rename old-api into functional api
