+++ src/g-key/functional-api.ts
