+++ src/lcd/functional-api.ts
