+++ src/lcd/index.ts
@@ -1,4 +1,4 @@
 export { LogiLcd } from './logi-lcd';
 
-import * as lcdTmp from './old-api';
+import * as lcdTmp from './functional-api';
 export const lcd = lcdTmp;
