+++ src/tests/g-key/run-tests.ts
@@ -1,6 +1,6 @@
 import { executeTest } from '../test-helper';
 import * as tests from './logi-g-key';
-import * as oldTests from './old-api';
+import * as functionalTests from './functional-api';
 
 
 export async function runGkeyTests()
@@ -8,12 +8,12 @@ export async function runGkeyTests()
 	console.log('Old API');
 	console.log('---');
 
-	await executeTest('pressed keys', () => oldTests.testPressedKeys(10e3));
-	oldTests.shutdown();
-	await executeTest('init with callback', () => oldTests.testCallback(10e3));
-	oldTests.shutdown();
-	await executeTest('init with callback and context', () => oldTests.testContext(10e3));
-	oldTests.shutdown();
+	await executeTest('pressed keys', () => functionalTests.testPressedKeys(10e3));
+	functionalTests.shutdown();
+	await executeTest('init with callback', () => functionalTests.testCallback(10e3));
+	functionalTests.shutdown();
+	await executeTest('init with callback and context', () => functionalTests.testContext(10e3));
+	functionalTests.shutdown();
 
 	console.log('');
 	console.log('New API');
