+++ src/tests/lcd/run-tests.ts
@@ -1,6 +1,6 @@
 import { executeTest } from '../test-helper';
 import * as tests from './logi-lcd';
-import * as oldTests from './old-api';
+import * as functionalTests from './functional-api';
 
 
 export async function runLcdTests()
@@ -8,9 +8,9 @@ export async function runLcdTests()
 	console.log('Old API');
 	console.log('---');
 
-	await executeTest('normal text', oldTests.testText);
-	await executeTest('random background', oldTests.testBackground);
-	oldTests.shutdown();
+	await executeTest('normal text', functionalTests.testText);
+	await executeTest('random background', functionalTests.testBackground);
+	functionalTests.shutdown();
 
 	console.log('');
 	console.log('New API');
